Corrigido
Corrigido.

+++ src/app/components/profile-page/project-card/project-card.component.ts
@@ -119,8 +119,8 @@ export class ProjectCardComponent implements OnInit, AfterViewInit {
     this.imageSubs.unsubscribe();
   }
 
-  deleteclick(post:any){
-    document.getElementById('deleteProject')?.setAttribute('data-project', this.project.postId);
+  deleteclick(){
+    document.getElementById('deleteProject')?.setAttribute('data-project', this.project.projectId);
     document.getElementById('deleteProject')?.setAttribute('data-profileId', this.project.profileId);
   }
 
@@ -132,19 +132,22 @@ export class ProjectCardComponent implements OnInit, AfterViewInit {
   }
 
   async onDelete(){
-    try{
-      const delprojectid = document.getElementById('deleteProject')?.getAttribute('data-project');
-      const delprofileid = document.getElementById('deleteProject')?.getAttribute('data-profileId');
-      if (delprofileid != null && delprojectid != null){
-      await this.projectsService.deleteProject(delprofileid, delprojectid);
-      this.projectDeleted.emit(delprojectid);
-    }
-    } catch (error){
-      console.error(error);
-    } finally {
-      let close = document.getElementById('close');
-      close?.click();
-    }
+    // If the user is the user, then he can delete it
+    if (this.profileId === this.userProfile) {
+      try{
+        const delprojectid = document.getElementById('deleteProject')?.getAttribute('data-project');
+        const delprofileid = document.getElementById('deleteProject')?.getAttribute('data-profileId');
+        if (delprofileid != null && delprojectid != null){
+        await this.projectsService.deleteProject(delprofileid, delprojectid);
+        this.projectDeleted.emit(delprojectid);
+      }
+      } catch (error){
+        console.error(error);
+      } finally {
+        let close = document.getElementById('close');
+        close?.click();
+      }
+    }    
   }
 
 }
